Making the program more modular
